+++ src/pages/auth/Register/Form.tsx
@@ -24,8 +24,8 @@ import {
   Theme
 } from '@material-ui/core/styles'
 import { withTranslate, TranslationContextProps, ReduxState } from 'ra-core'
-import { extractError } from './errorSagas'
-import { crudCreate } from './crudCreate'
+import { extractError } from '../../../errorSagas'
+import { crudCreate } from '../../../crudCreate'
 
 interface Props {
   redirectTo?: string
