+++ src/pages/index.tsx
@@ -1,0 +1,8 @@
+import Register from './auth/Register'
+import Login from './auth/Login'
+import Notify from './auth/Notify'
+import PasswordReset from './auth/PasswordReset'
+import PasswordSet from './auth/PasswordSet'
+import ResendPage from './auth/ResendPage'
+
+export { ResendPage, PasswordSet, PasswordReset, Notify, Login, Register }
